refactoring du cas de la gestion multi-fichiers: MediaType est deplace dans Form/type/ les fichiers Entity/Core/Media et MediaLink sont supprimés ainsi que Form/EmbedTypes/MediaLinkEmbedType.php
